requirements.txt
numpy
pandas
flask
flask-cors
newspaper3k
sklearn
nltk
tensorflow
matplotlib
scikit-plot
textblob
gunicorn
validators
newsapi-python

flask-login
flask-sqlalchemy
flask-mysqldb
flask-wtf
Werkzeug